client

+++ Server/app/ServerView/viewComment/views.py
@@ -32,7 +32,9 @@ def get_Comments(articleid):
         else:
             entities[fid].setdefault('soncomment', []).append(entitiy)
     #取前N个
-    return jsonify(Common.trueReturn(l[(int(pageNumber)-1)*int(pageSize):int(pageNumber)*int(pageSize)],'query ok'))
+    left=min(len(l),max(0,(int(pageNumber)-1)*int(pageSize)))
+    right=min(len(l),max(0,int(pageNumber)*int(pageSize)))
+    return jsonify(Common.trueReturn(l[left:right],'query ok'))
 
 @comment_blue.route("/",methods=["POST"])
 @IdentifyUtil.login_required
@@ -42,13 +44,35 @@ def post_Comments():
     if not userid :
         return jsonify(Common.falseReturn(None,'login required'))
     params = request.get_json()
+    if 'articleid' not in params or 'comment' not in params:
+        return jsonify(Common.falseReturn('articleid and comment is required'))
     res = CommentApi.postComment(params.get('articleid'),userid,params.get('comment'),params.get('refid'))
     return jsonify(res)
 
 @comment_blue.route("/<commentid>",methods=["DELETE"])
 @IdentifyUtil.login_required
 def delete_Comments(commentid):
-    return jsonify(CommentApi.deleteComment(commentid))
+    #获取自己id，只能删除自己文章下或者自己的评论
+    userid = IdentifyUtil.get_user_id()
+    if not userid :
+        return jsonify(Common.falseReturn(None,'login required'))
+    if CommentApi.getIsSelfComment(userid,commentid)['data'] or CommentApi.getIsSelfArticlesComment(userid,commentid)['data']:
+        return jsonify(CommentApi.deleteComment(commentid))
+    return jsonify(Common.falseReturn(None,'permission denied'))
+
+@comment_blue.route("/<commentid>",methods=["PUT"])
+@IdentifyUtil.login_required
+def update_Comments(commentid):
+    #获取自己id，只能修改自己的评论
+    userid = IdentifyUtil.get_user_id()
+    if not userid :
+        return jsonify(Common.falseReturn(None,'login required'))
+    params = request.get_json()
+    if 'comment' not in params:
+        return jsonify(Common.falseReturn(None,'comment is required'))
+    if not CommentApi.getIsSelfComment(userid,commentid)['data']:
+        return jsonify(Common.falseReturn(None,'permission denied'))
+    return jsonify(CommentApi.updateCommentByCommentId(commentid,params.get('comment')))
 
 @comment_blue.route("/counts/topcomments/<articleid>",methods=["GET"])
 def get_articleCommentCounts(articleid):
@@ -57,3 +81,12 @@ def get_articleCommentCounts(articleid):
 @comment_blue.route("/counts/childcomments/<commentid>",methods=["GET"])
 def get_childCommentCounts(commentid):
     return jsonify(CommentApi.getChildCommentCountByCommentId(commentid))
+
+
+@comment_blue.route("/belongs/<commentid>",methods=["GET"])
+@IdentifyUtil.login_required
+def get_IsSelfComment(commentid):
+    userid = IdentifyUtil.get_user_id()
+    if not userid:
+        return jsonify(Common.falseReturn(None, 'login required'))
+    return jsonify(CommentApi.getIsSelfComment(userid,commentid))
